✨ Add basic /buy command
Currently, items are not actually purchased. The total cost is taken from the user's cash based on the item's price and the amount they buy, but nothing is added to their inventory yet. Also, the buy menu hasn't been implemented yet so nothing happens if no item code is given.